+++ Website/src/components/Markdown/index.tsx
@@ -1,17 +1,18 @@
 import Markdown from "markdown-to-jsx";
-import Anchor from "../components/dapi/Anchor";
-import Video from "../components/dapi/Video";
-import Audio from "../components/dapi/Audio";
-import Checkmark from "../components/icons/Checkmark";
-import Dangermark from "../components/icons/Dangermark";
-import Warnmark from "../components/icons/Warnmark";
+import Anchor from "../dapi/Anchor";
+import Video from "../dapi/Video";
+import Audio from "../dapi/Audio";
+import Checkmark from "../icons/Checkmark";
+import Dangermark from "../icons/Dangermark";
+import Warnmark from "../icons/Warnmark";
 import Icon from "@mui/material/Icon";
 import Typography from "@mui/material/Typography";
 import React from "react";
-import Image from "./dapi/Image";
-import { Box, Chip, Container, Divider, Grid, Paper, Stack } from "@mui/material";
-import { StyledMarkdown } from "../styles/StyledMarkdon";
+import Image from "../dapi/Image";
+import { Alert, AlertTitle, Box, Chip, Container, Divider, Grid, Paper, Stack } from "@mui/material";
+import { StyledMarkdown } from "../../styles/StyledMarkdon";
 import styled from "@emotion/styled";
+import { parser } from "./parser";
 
 type Props = {
   children: string;
@@ -24,7 +25,17 @@ export function Markup(props: Props) {
     "h1, & h2, & h3, & h4, & h5, & h6": {
       border: "none",
     },
+    "& .MuiDivider-fullWidth": {
+      "& .MuiDivider-root::before, & .MuiDivider-root::after": {
+        borderTop: "1px solid hsl(210, 18.2%, 87.1%)",
+      },
+    },
   });
+
+  const StyledAlert = styled(Alert)((props) => ({
+    marginTop: 4,
+    marginBottom: 4,
+  }));
 
   return (
     <StyledMarkdown style={{ display: "inline-block", padding: "8px", height: "100%", width: "100%", ...props.style }}>
@@ -84,7 +95,7 @@ export function Markup(props: Props) {
             },
           },
         }}
-        children={props.children}
+        children={parser(props.children)}
       />
     </StyledMarkdown>
   );