+++ Website/src/components/Markdown/parser.tsx
@@ -1,0 +1,8 @@
+import { rules } from "./rules";
+
+export function parser(text: string) {
+  rules.forEach(([rule, template]) => {
+    text = text.replace(rule, template as string);
+  });
+  return text;
+}