+++ Website/src/components/Material3.ts
@@ -146,7 +146,7 @@ const Material3 = {
       letterSpacing: "0",
       verticalAlign: "middle",
       backgroundColor: theme.palette.primary.main,
-      borderRadius: "8px",
+      // borderRadius: "8px",
       transition: "all 0.25s linear",
       boxShadow:
         "0 2px 2px 0 rgba(0, 0, 0, 0.14), 0 1px 5px 0 rgba(0, 0, 0, 0.12),\r\n    0 3px 1px -2px rgba(0, 0, 0, 0.2)",