+++ CHANGELOG.md
@@ -9,3 +9,4 @@
 - Make edit groups better
 - Added support to delete single cards
 - Using now the powerfull api of WSP
+- Improved markdown formatting